fix: tiny change

+++ test/cpv/library-checker-linear_algebra/sparse_matrix_det.mints-ms64.cpp
@@ -24,7 +24,7 @@ int main() {
     a(x, y) = z;
   }
   auto is0 = [](cT_(mint) x) { return x.val() == 0; };
-  tifa_libs::rand_gen<u32> gen(1, mint::mod() - 1);
+  tifa_libs::rand_gen<u32> gen(1, u32((i32)mint::mod() - 1));
   std::cout << tifa_libs::det_rd(a, gen, is0) << '\n';
   return 0;
 }
